# mail-worker/doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cloud-Mail 个人 API 系统",
    "description": "Cloud-Mail 的公开 API 接口文档，使用 API Key 进行认证",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-domain.com",
      "description": "生产服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "所有 /v1 API 请求都需要在请求头中包含有效的 API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        }
      },
      "Email": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "邮件ID"
          },
          "from": {
            "type": "string",
            "description": "发件人邮箱"
          },
          "subject": {
            "type": "string",
            "description": "邮件主题"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "邮件创建时间"
          },
          "text": {
            "type": "string",
            "description": "邮件纯文本内容"
          },
          "html": {
            "type": "string",
            "description": "邮件HTML内容"
          },
          "address": {
            "type": "string",
            "description": "收件邮箱地址"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "description": "总项目数"
          },
          "totalPages": {
            "type": "integer",
            "description": "总页数"
          },
          "currentPage": {
            "type": "integer",
            "description": "当前页码"
          },
          "limit": {
            "type": "integer",
            "description": "每页项目数"
          }
        }
      },
      "EmailListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "GenerateEmailRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "邮箱前缀（可选）"
          },
          "domain": {
            "type": "string",
            "description": "邮箱域名（可选）"
          }
        }
      },
      "GenerateEmailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "邮箱ID"
          },
          "address": {
            "type": "string",
            "description": "完整的邮箱地址"
          }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "邮箱账户ID"
          },
          "address": {
            "type": "string",
            "description": "邮箱地址"
          },
          "name": {
            "type": "string",
            "description": "自定义名称"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "AccountListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/v1/emails": {
      "get": {
        "summary": "获取当前用户的所有邮箱列表（分页）",
        "description": "返回通过此 API Key 认证的用户所拥有的全部邮箱地址列表",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码（从1开始）"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 25
            },
            "description": "每页项目数"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/generate": {
      "post": {
        "summary": "创建临时邮箱",
        "description": "为认证用户创建一个新的临时邮箱",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateEmailRequest"
              },
              "example": {
                "name": "test",
                "domain": "example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateEmailResponse"
                },
                "example": {
                  "id": 123,
                  "address": "test@example.com"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/{emailAddress}/messages": {
      "get": {
        "summary": "获取邮件列表（分页）",
        "description": "获取指定邮箱地址的邮件列表，支持分页",
        "parameters": [
          {
            "in": "path",
            "name": "emailAddress",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "邮箱地址"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码（从1开始）"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 25
            },
            "description": "每页项目数"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailListResponse"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "邮箱地址不存在或无访问权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/{emailAddress}/messages/{messageId}": {
      "get": {
        "summary": "获取单封邮件详情",
        "description": "获取指定邮件的完整详细信息",
        "parameters": [
          {
            "in": "path",
            "name": "emailAddress",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "邮箱地址"
          },
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "邮件ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "邮件不存在或无访问权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}